refactor: 전역 타입 app -> shared로 이동
