+++ src/refactoring/features/cart/ui/CartItem.tsx
@@ -1,4 +1,4 @@
-import { ICartItem } from '../../../app/types';
+import { ICartItem } from '../../../shared/types';
 
 interface CartItemProps {
   item: ICartItem;
